Start splitting renew.py out of cli.py
